requirements.txt
backcall==0.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jupyter_core==5.3.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.2
python-pptx==0.6.23
pyzmq==25.0.2
six==1.16.0
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.12.2
wcwidth==0.2.6
Werkzeug==3.0.3
XlsxWriter==3.2.0
